requirements.txt
Django==2.2
djangorestframework==3.9.2
djangorestframework-filters==0.10.2.post0
jsonify==0.5
pgadmin4==2.0
requests==2.21.0
virtualenv==16.5.0
